requirements.txt
PyYAML~=6.0
numpy~=1.23.4
tqdm~=4.64.1
scipy~=1.9.3
torch~=1.12.1
matplotlib~=3.6.1
Pillow~=9.2.0